flowmamba_xbd_pixel_tau_corn_upper/requirements.txt
torch
torchvision
PyYAML
Pillow
numpy
shapely

# Optional but recommended for real upstream FlowMamba / VMamba backbone loading:
timm
einops
mmengine
fvcore
